UI looks good, elos are calculated
