+++ backend/data/ping_pong.csv
@@ -1,5 +1,4 @@
 name,elo,wins,losses,last10
-Joona,1000,0,10,llllllllll
-Tommi,1000,14,8,wwwwwwwwwl
-Mika,1000,9,4,wwwwllllww
-Mikko,1000,0,1,l
+Tommi,1036,10,3,wwwwwwwlll
+Mika,967,4,10,llllllwwww
+Joona,997,3,4,llllwww
