+++ test_percentages.py
@@ -1,5 +1,5 @@
 import random
-from elocalculator import Player, get_new_elos,set_new_elos
+from backend.elocalculator import Player, set_new_elos
